requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib
bcrypt == 4.0.1
python-dotenv
pytest
httpx
sqlmodel
python-multipart